src/requirements.txt
python-dotenv==1.0.1
httpx==0.27.0

--extra-index-url https://test.pypi.org/simple/
nitro-toolkit~=0.0.9-dev1